+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 
 ### Changed
 - Refactored methods from Runner into BaseRunner (#193)
+- Updated unit conversion constants to be consistent with ORCA (#269).
 
 ### Deprecated
 ### Removed